S-33 Make PIN code login safer by allowing max 3 pin codes in 15 min. Add logging of authentication events.

+++ src/api/auth.ts
@@ -30,24 +30,69 @@ authRoute.get(
 )
 
 authRoute.post('/api/auth/email/create-pin', rateLimit(5, 10), async (req, res, next) => {
-  const body = req.body
-  console.log(body)
-  const pin = Math.floor(100000 + Math.random() * 900000)
-  console.log('pin', pin)
-  await prisma.emailLoginPinCode.deleteMany({
-    where: {
-      email: body.email,
-    },
-  })
-  await prisma.emailLoginPinCode.create({
-    data: {
-      email: body.email,
-      pinCode: pin.toString(),
-    },
-  })
+  try {
+    const body = req.body
+    console.log(body)
 
-  console.log(`pin ${pin} sent to email ${body.email}`)
-  res.status(200).json({ message: 'OK' })
+    await prisma.log.create({
+      data: {
+        data: { email: body.email, ip: req.ip },
+        event: 'pin-requested',
+      },
+    })
+
+    // Check for existing pin codes
+    const existingPinCodes = await prisma.emailLoginPinCode.findMany({
+      where: {
+        email: body.email,
+        createdAt: {
+          gte: new Date(Date.now() - 15 * 60 * 1000) // 15 minutes ago
+        }
+      },
+    })
+
+    // If there are too many existing pin codes or recent attempts, reject
+    if (existingPinCodes.length >= 3) {
+      console.error(`Too many pin code attempts for ${body.email}. Please try again later.`)
+      return res.status(429).json({
+        message: 'Too many pin code attempts. Please try again later.'
+      })
+    }
+
+    // Generate new pin
+    const pin = Math.floor(100000 + Math.random() * 900000)
+    console.log('pin', pin)
+
+    // Delete pins older than 15 minutes
+    await prisma.emailLoginPinCode.deleteMany({
+      where: {
+        email: body.email,
+        createdAt: {
+          lt: new Date(Date.now() - 15 * 60 * 1000) // 15 minutes ago
+        }
+      },
+    })
+
+    // Create new pin
+    await prisma.emailLoginPinCode.create({
+      data: {
+        email: body.email,
+        pinCode: pin.toString(),
+      },
+    })
+
+    await prisma.log.create({
+      data: {
+        data: { email: body.email, ip: req.ip },
+        event: 'pin-created',
+      },
+    })
+
+    console.log(`pin ${pin} sent to email ${body.email}`)
+    res.status(200).json({ message: 'OK' })
+  } catch (error) {
+    next(error)
+  }
 })
 
 authRoute.post('/api/auth/email/check-pin', rateLimit(10, 1), passport.authenticate('local'), (req, res, next) => {
